requirements.txt
pyserial 
datetime 
dnspython
requests
pymongo
schedule
